requirements.txt
PyGithub

requests
configparser
pprint